add all pre-handle data
