+++ camel/loaders/autoconvert_dataformat.py
@@ -11,34 +11,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
-r"""Module for processing and transforming JSON/JSONL data files.
 
-This module provides functions to load, transform, and save JSON/JSONL data
-files. It supports generating mapping rules between source and target JSON
-formats, and applying these rules to transform data accordingly. The module
-also includes utilities for cleaning JSON responses and handling nested
-structures.
 
-Functions:
-    load_json: Load data from a single JSON or JSONL file.
-    load_multiple_json: Load data from multiple JSON or JSONL files.
-    get_all_keys: Recursively extract all keys from nested JSON structures.
-    clean_json_response: Clean markdown formatting from AI-generated JSON.
-    generate_mapping_rules: Generate mapping rules between source and target JSON.
-    transform_data: Transform source data using mapping rules.
-    save_json: Save data to a JSON file.
-    main: Main function to process JSON/JSONL files for data transformation.
-"""
-
+import argparse
 import json
 import os
-from camel.agents import ChatAgent
 
+from camel.agents import ChatAgent
 from camel.configs import ChatGPTConfig
 from camel.models import ModelFactory
 from camel.types import ModelPlatformType, ModelType
-import argparse
-
 
 model = ModelFactory.create(
     model_platform=ModelPlatformType.OPENAI,
@@ -50,6 +32,7 @@ sys_msg = "You are a helpful assistant."
 
 # Initialize ChatAgent
 camel_agent = ChatAgent(system_message=sys_msg, model=model)
+
 
 def load_json(file_path, n=None):
     r"""Read data from a JSON or JSONL file.
@@ -75,7 +58,9 @@ def load_json(file_path, n=None):
                 file_path = alt_path
                 break
         else:
-            raise FileNotFoundError(f"No JSON or JSONL file found for {file_path}")
+            raise FileNotFoundError(
+                f"No JSON or JSONL file found for {file_path}"
+            )
 
     with open(file_path, 'r') as file:
         # Check the file extension
@@ -95,6 +80,7 @@ def load_json(file_path, n=None):
             else:
                 return [data]
 
+
 def load_multiple_json(file_paths, n=None):
     r"""Read data from multiple JSON or JSONL files.
 
@@ -109,17 +95,18 @@ def load_multiple_json(file_paths, n=None):
     """
     if isinstance(file_paths, str):
         file_paths = [file_paths]
-    
+
     all_data = []
     for file_path in file_paths:
         try:
             data = load_json(file_path, n)
             all_data.extend(data if isinstance(data, list) else [data])
         except Exception as e:
-            print(f"Error processing file {file_path}: {str(e)}")
+            print(f"Error processing file {file_path}: {e!s}")
             continue
-    
+
     return all_data
+
 
 def get_all_keys(data, prefix=''):
     r"""Recursively get all keys from a nested JSON structure.
@@ -138,9 +125,10 @@ def get_all_keys(data, prefix=''):
             keys.add(current_key)
             keys.update(get_all_keys(value, f"{current_key}."))
     elif isinstance(data, list):
-        for i, item in enumerate(data):
-            keys.update(get_all_keys(item, f"{prefix}[{i}]."))
+        for _, item in enumerate(data):
+            keys.update(get_all_keys(item, f"{prefix}[{_}]."))
     return keys
+
 
 def clean_json_response(response):
     r"""Clean markdown formatting from AI-generated JSON content.
@@ -152,7 +140,7 @@ def clean_json_response(response):
         str: The cleaned JSON content without markdown formatting.
     """
     content = response.strip()
-    
+
     # Remove markdown code block markers
     if content.startswith('```'):
         # Find the positions of the first and last ```
@@ -160,8 +148,9 @@ def clean_json_response(response):
         last = content.rfind('```')
         if first != -1 and last != -1:
             content = content[first:last].strip()
-    
+
     return content
+
 
 def generate_mapping_rules(source_data, target_data):
     r"""Generate mapping rules between source and target JSON structures.
@@ -178,9 +167,8 @@ def generate_mapping_rules(source_data, target_data):
         json.JSONDecodeError: If the AI response is not valid JSON.
     """
     # Get all keys from source and target data
-    source_keys = get_all_keys(source_data)
     target_keys = get_all_keys(target_data)
-    
+
     # Build the prompt
     prompt = f"""
     Here are two JSON data samples:
@@ -191,9 +179,10 @@ def generate_mapping_rules(source_data, target_data):
     Target JSON:
     {json.dumps(target_data, indent=5)}
 
-    Target JSON has the following keys: {sorted(list(target_keys))}
+    Target JSON has the following keys: {sorted(target_keys)}
 
-    Please analyze the keys in the source JSON and target JSON, and generate a mapping rule.
+    Please analyze the keys in the source JSON and target JSON, 
+    and generate a mapping rule.
     IMPORTANT: 
     1. Your response must be ONLY a valid JSON object.
     2. Only map to keys that exist in the target JSON structure.
@@ -212,20 +201,27 @@ def generate_mapping_rules(source_data, target_data):
         # Clean the response content
         cleaned_content = clean_json_response(response.msgs[0].content)
         mapping = json.loads(cleaned_content)
-        
-        # Verify that the mapping rules only contain keys existing in the target format
+
+        # Verify that the mapping rules only contain keys
+        #  existing in the target format
         invalid_target_keys = set(mapping.values()) - target_keys
         if invalid_target_keys:
-            raise ValueError(f"Mapping contains invalid target keys: {invalid_target_keys}")
-            
+            raise ValueError(
+                f"Mapping contains invalid target keys: {invalid_target_keys}"
+            )
+
         return mapping
-    except json.JSONDecodeError as e:
-        print(f"Error: AI response was not valid JSON. Response content: {response.msgs[0].content}")
+    except json.JSONDecodeError:
+        print(
+            f"""Error: AI response was not valid JSON. 
+            Response content: {response.msgs[0].content}"""
+        )
         print(f"Cleaned content was: {cleaned_content}")
         raise
     except Exception as e:
-        print(f"Error generating mapping rules: {str(e)}")
+        print(f"Error generating mapping rules: {e!s}")
         raise
+
 
 def transform_data(source_data, mapping_rules, target_structure):
     r"""Transform source data using mapping rules.
@@ -239,18 +235,19 @@ def transform_data(source_data, mapping_rules, target_structure):
         dict: Transformed JSON data matching the target structure.
     """
     transformed_data = {}
-    
+
     # Only transform fields specified in the mapping rules
     for source_key, target_key in mapping_rules.items():
         if source_key in source_data:
             # Handle nested keys
             keys = target_key.split('.')
             current_dict = transformed_data
-            for i, key in enumerate(keys[:-1]):
+            for _, key in enumerate(keys[:-1]):
                 current_dict = current_dict.setdefault(key, {})
             current_dict[keys[-1]] = source_data[source_key]
-    
+
     return transformed_data
+
 
 def save_json(data, file_path):
     r"""Save JSON data to a file.
@@ -262,19 +259,27 @@ def save_json(data, file_path):
     with open(file_path, 'w') as file:
         json.dump(data, file, indent=4)
 
+
 def main():
     r"""Main function to process JSON/JSONL files for data transformation.
 
     This function handles command-line arguments, loads source and target data,
     generates mapping rules, transforms the data, and saves the output.
     """
-    import argparse
-    
-    parser = argparse.ArgumentParser(description='Process JSON/JSONL files for data transformation')
-    parser.add_argument('input_files', nargs='+', help='Input JSON/JSONL file(s)')
-    parser.add_argument('--target', required=True, help='Target JSON file for format reference')
+
+    parser = argparse.ArgumentParser(
+        description='Process JSON/JSONL files for data transformation'
+    )
+    parser.add_argument(
+        'input_files', nargs='+', help='Input JSON/JSONL file(s)'
+    )
+    parser.add_argument(
+        '--target', required=True, help='Target JSON file for format reference'
+    )
     parser.add_argument('--output', required=True, help='Output JSON file')
-    parser.add_argument('-n', type=int, help='Number of lines to process from each input file')
+    parser.add_argument(
+        '-n', type=int, help='Number of lines to process from each input file'
+    )
     args = parser.parse_args()
 
     # Load source data from multiple files
@@ -288,7 +293,7 @@ def main():
 
     # Generate mapping rules
     mapping_rules = generate_mapping_rules(source_data[0], target_data)
-    
+
     # Transform all data
     transformed_data = []
     for item in source_data:
@@ -296,13 +301,17 @@ def main():
             transformed = transform_data(item, mapping_rules, target_data)
             transformed_data.append(transformed)
         except Exception as e:
-            print(f"Error transforming item: {str(e)}")
+            print(f"Error transforming item: {e!s}")
             continue
 
     # Save transformed data
     save_json(transformed_data, args.output)
-    print(f"Processed {len(transformed_data)} items from {len(args.input_files)} files")
+    print(
+        f"""Processed {len(transformed_data)} 
+        items from {len(args.input_files)} files"""
+    )
     print(f"Output saved to: {args.output}")
+
 
 if __name__ == "__main__":
     main()
