+++ examples/dataset/reasoning_maths/download_huggingface_data.py
@@ -1,3 +1,16 @@
+# ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ========= Copyright 2023-2024 @ CAMEL-AI.org. All Rights Reserved. =========
 from datasets import load_dataset
 
 # Load dataset
@@ -7,4 +20,4 @@ print(ds)
 # Export the dataset to a JSON file
 ds["test"].to_json("./downloadeddata_files/HuggingFaceH4_MATH-500.json")
 
-print("Dataset has been saved as a JSON file: ./downloadeddata_files/HuggingFaceH4_MATH-500.json")
+print("Dataset has been saved as a JSON file")
